realize the 3d floodfill selection
rename the folder name

Revert "3dfloodfill with Cython"

test code for ray cursor

get the ray cursor line

Create points_selection_vispy.py

Lasso & shape selection and picking for 3d points data, controlled by key 1-4, the bug of empty the drawing line visual should be fixed later.

Create points_selection_vispy.py

Lasso & shape selection and picking for 3d points data, controlled by key 1-4, the bug of empty the drawing line visual should be fixed later.

Revert "Create points_selection_vispy.py"

enable the picking for volume and get the max pos along eye sight
