+++ flood_fill/zoo_test.jpg
