--- floodfill_selection/multivol/__init__.py
@@ -1,3 +1,0 @@
-from .rgba_volume_visual import RGBAVolume
-from .multi_volume_visual import MultiVolume
-from .colors import get_translucent_cmap
